[+] Add fd_set_manage server function and fix CI/CD
