+++ .github/workflows/cmake.yml
@@ -36,4 +36,4 @@ jobs:
       working-directory: ${{github.workspace}}/build
       # Execute tests defined by the CMake configuration.
       # See https://cmake.org/cmake/help/latest/manual/ctest.1.html for more detail
-      run: make tests_run
+      run: make tests
