+++ Makefile
@@ -24,7 +24,13 @@ all:
 	cmake . && make
 
 tests:
-	cmake --build ./build
-	cd ./build ; ctest
+	cmake -B ./build-with-coverage -G  "Unix Makefiles" -DCMAKE_BUILD_TYPE=Debug -DENABLE_CODE_COVERAGE=ON
+	cmake --build ./build-with-coverage
+	cd ./build-with-coverage ; ctest
+	mkdir coverage-report
+	find . -name "*.gcda" -exec cp {} coverage-report/ \;
+	find . -name "*.gcno" -exec cp {} coverage-report \;
+	cd coverage-report
+	gcovr -r ../../
 
 .PHONY: all clean fclean re tests
