+++ server/net_utils/fd_sets_manage/server_accept_new_client.c
@@ -1,0 +1,13 @@
+/*
+** EPITECH PROJECT, 2022
+** mes_equipes
+** File description:
+** server_accept_new_client
+*/
+
+#include "net_utils.h"
+
+void server_accept_new_client(tcp_server_t *srv)
+{
+    (void)srv;
+}
