+++ server/net_utils/fd_sets_manage/server_write_client.c
@@ -1,0 +1,13 @@
+/*
+** EPITECH PROJECT, 2022
+** mes_equipes
+** File description:
+** server_write_client
+*/
+
+#include "net_utils.h"
+
+void server_write_client(tcp_server_t *srv, peer_t *tmp) {
+    (void) srv;
+    (void) tmp;
+}
